Warn when symbols are present with a V2 source specified
If symbol packages are found during the push command that would normally be uploaded to a symbol server. Warn the user if the `source` specified would not be considered a `V3` source.

If the `source` is not considered `V3` then the symbol package upload will be aborted in a different section of the existing code.